modifies 'Challenge.Application' - add new domain events to proccess financial transactions before sale or purchase of a product

+++ src/XPChallenge.Application/Features/FinancialProduct/Commands/CreateProductCommand.cs
@@ -5,7 +5,7 @@ namespace XPChallenge.Application.Features.FinancialProduct.Commands;
 public class CreateProductCommand : IRequest<CreateProductResponse>
 {
     public string Name { get; set; }
-    public decimal CurrentPurchasePrice { get; set; }
+    public double CurrentPurchasePrice { get; set; }
     public DateTime DueDate { get; set; }
 
     public Domain.Entities.FinancialProduct MapToEntity()
